requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
websockets==12.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0 